tech-digest/requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
requests==2.32.3
twilio==9.3.1
PyYAML==6.0.2
